+++ doc/versionHistory.rst
@@ -13,6 +13,7 @@ Version History
 -------------
 
 * Change intrinsic map interpolation to LinearNDInterpolator to handle irregular grid of intrinsic Zernike measurements that appear due to vignetting in the model that creates the maps.
+* Add small test for intrinsic map creation in test_calcZernikesDanishTaskCwfs.py.
 
 .. _lsst.ts.wep-16.0.1: